Add files via upload
I have made a game about Harry Potter who is flying on a broom, and he has to avoid dementors coming for the right side of the screen. By using the arrow keys you can Harry around.
When starting the game you must press the green flag twice in order for everything to work!